+++ src/ui/color_picker_gray.rs
@@ -2,8 +2,8 @@ use crate::ui::utils::*;
 use crate::AppState;
 use little_tui::*;
 
-static X: i16 = 16;
-static Y: i16 = 2;
+static X: isize = 16;
+static Y: isize = 2;
 
 pub fn build() -> Element<AppState> {
     let color_picker_gray: Element<AppState> = Element::new();
