+++ src/ui/screen.rs
@@ -11,8 +11,8 @@ pub fn build() -> Element<AppState> {
 
     screen
         .on_loop(|_el, state, _event| {
-            let x = (Terminal::columns().saturating_sub(APP_WIDTH) / 2) as i16;
-            let y = (Terminal::rows().saturating_sub(APP_HEIGHT) / 2) as i16;
+            let x = (Terminal::columns().saturating_sub(APP_WIDTH) / 2) as isize;
+            let y = (Terminal::rows().saturating_sub(APP_HEIGHT) / 2) as isize;
 
             if x != state.app_x || y != state.app_y {
                 Terminal::clear_screen();
