+++ src/ui/utils.rs
@@ -86,8 +86,8 @@ pub(crate) fn canvas_data_from_click(
     size: usize,
     data: &mut Vec<Option<u8>>,
     paintbrush: Option<u8>,
-    mouse_x: i16,
-    mouse_y: i16,
+    mouse_x: isize,
+    mouse_y: isize,
     fill: bool,
 ) {
     let row = mouse_y.saturating_sub(el.visual.y.get()) as usize;
